LeafDevice/requirements.txt
asyncio
python-dotenv
azure-iot-device
